#9 feat/ Added func7 for LIC6
As per description checks if there are at least 3 points then for every set of N_PTS consequtive points checks if there is a point that is futher away than DIST from the line created from the first and last point.

Simply check distance if both points are the exact same.

No tests yet so input and output has been commented out to be added later